pyproject.toml
[project]
name = "cfc-dc-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.7.1",
    "fastapi>=0.135.3",
    "jishaku>=2.6.3",
    "motor>=3.7.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
